Fix naming of ambiguous libs (#1625)
Currently, we emit symlinks that disambiguate native libs in the form of `libfoo.pic-{hash}.a` when building shared libraries in `opt` mode. I am not sure if anything depends on these output files to end with `.pic.a` instead, however I beileve it `libfoo-{hash}.pic.a` is a better name.