Revert #1564 (#1663)
* Add example workspace cross-compiling with zig

We've run into a number of issues, both in rustc and in cargo build
scripts, where we're not properly propagating cc_toolchain information
to compiles and particularly link actions. These have often been hard to
test because we don't have an example set-up with a non-default C++
toolchain.

By wiring up a zig cross-compiling toolchain, we have a concrete example
we can work with. It's also an interesting potential example for folks
who want to set something similar up.

* Revert #1564

It looks like the reported issue is probably related to the ios
toolchain in use, rather than our general toolchain flag propagation.

* Add platform_mappings and --cpu flag to ios_build example