Improve handling of build.rs link dependencies (#1970)
This PR helps clarify the difference between cargo_build_script build
dependencies, and link dependencies. It separates the two, and ensures
that link_deps are only for direct (non-transitive) normal dependencies
of the build crate. It does not fix the issue with how crates universe
detects links crates although it does add a comment explaining the
issue.